add new class ChordGraphics and move drawing logic
