+++ ChordImages/ChordImages.csproj
@@ -45,6 +45,7 @@
     <Compile Include="Chord.cs" />
     <Compile Include="ChordBoxImage.cs" />
     <Compile Include="Fingers.cs" />
+    <Compile Include="ChordGraphics.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
